refactor faer module logic and factor out faer tests
