+++ crates/argmin-math/src/faer_m_0_20/div.rs
@@ -92,8 +92,8 @@ impl<E: Entity + Div<E, Output = E>> ArgminDiv<Mat<E>, Mat<E>> for Mat<E> {
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_helper::*;
     use super::*;
+    use crate::faer_tests::test_helper::*;
     use approx::assert_relative_eq;
     use faer::mat::AsMatRef;
     use paste::item;
