+++ crates/argmin-math/src/faer_m_0_20/dot.rs
@@ -158,8 +158,8 @@ mod multiply_matrix_with_scalar {
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_helper::*;
     use super::*;
+    use crate::faer_tests::test_helper::*;
     use approx::assert_relative_eq;
     use faer::mat::AsMatRef;
     use paste::item;
