+++ crates/argmin-math/src/faer_m_0_20/l1norm.rs
@@ -15,8 +15,8 @@ impl<E: Entity + ComplexField> ArgminL1Norm<E::Real> for Mat<E> {
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_helper::*;
     use super::*;
+    use crate::faer_tests::test_helper::*;
     use approx::assert_relative_eq;
     use paste::item;
