+++ crates/argmin-math/src/faer_m_0_20/mod.rs
@@ -48,6 +48,3 @@ pub use signum::*;
 pub use sub::*;
 pub use transpose::*;
 pub use zero::*;
-
-#[cfg(test)]
-mod test_helper;
