+++ crates/argmin-math/src/faer_m_0_20/mul.rs
@@ -90,8 +90,8 @@ impl<E: SimpleEntity + ComplexField> ArgminMul<Mat<E>, Mat<E>> for Mat<E> {
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_helper::*;
     use super::*;
+    use crate::faer_tests::test_helper::*;
     use approx::assert_relative_eq;
     use faer::mat::AsMatRef;
     use faer::Mat;
