+++ crates/argmin-math/src/faer_m_0_20/signum.rs
@@ -60,8 +60,8 @@ where
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_helper::*;
     use super::*;
+    use crate::faer_tests::test_helper::*;
     use approx::assert_relative_eq;
     use faer::mat;
     use faer::Mat;
