+++ crates/argmin-math/src/faer_tests.rs
@@ -1,0 +1,2 @@
+mod add;
+pub mod test_helper;
